+++ src/utils.js
@@ -239,6 +239,10 @@ export const getNextSelectedRowsState = (selectedRows, row) => {
   return newSelectedRows;
 };
 
+export const extractPickSlipRequestIds = (pickSlipsData) => {
+  return pickSlipsData.map(pickSlip => pickSlip?.request?.requestID);
+}
+
 export const convertToSlipData = (source, intl, timeZone, locale, slipName = SLIPS_TYPE.PICK_SLIP) => {
   return source.map(pickSlip => {
     const {